% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{files_by_year}
\alias{files_by_year}
\title{It creates csv by year}
\usage{
files_by_year(path)
}
\arguments{
\item{path}{[\code{character}]\cr
path where the files are}
}
\description{
It aggregates and creates csv files by years.
It will save the results in folder results. This folder results
always will be created, if already exists, it will be deleted and after will be created
}
