% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing data.table}
\usage{
preprocessing(dt, colnames)
}
\arguments{
\item{dt}{[\code{data.table}]\cr
data.table which will be transformed}
}
\description{
This function performs the follow
data.table transformations:
\itemize{
\item Replace (,) with (.).
\item Changes variable type to numeric type
}
}
\examples{
 \dontrun{

}

}
